Pagination test.
Also refactoring code.
