+++ testai/src/main/java/adminPages/EventLogPage.java
@@ -1,0 +1,23 @@
+package adminPages;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import pages.AbstractObjectPage;
+
+public class EventLogPage extends AbstractObjectPage {
+
+    // input field
+    @FindBy(tagName = "input")
+    public WebElement inputPageLink;
+
+
+    public void enterPageNumber(String pageNumber) {
+        inputPageLink.sendKeys(pageNumber);
+    }
+
+    // constructor
+    public EventLogPage(WebDriver driver) {
+        super(driver);
+    }
+}
