+++ testai/src/test/java/basetest/BaseApiTest.java
@@ -17,7 +17,7 @@ public class BaseApiTest {
     public static void setUp() {
 
         reqSpec = new RequestSpecBuilder().
-                setBaseUri("https://sextet.akademijait.vtmc.lt/test-darzelis/").
+                setBaseUri("https://sextet.akademijait.vtmc.lt/darzelis/").
                 setContentType(ContentType.JSON).
                 addFilters(Arrays.asList(new RequestLoggingFilter(), new ResponseLoggingFilter())).
                 build();
