+++ testai/src/test/java/basetest/BaseTest.java
@@ -16,7 +16,7 @@ public class BaseTest {
         WebDriverManager.chromedriver().setup();
         driver = new ChromeDriver();
         driver.manage().window().maximize();
-        driver.get("https://sextet.akademijait.vtmc.lt/test-darzelis/");
+        driver.get("https://sextet.akademijait.vtmc.lt/darzelis/");
     }
 
     @AfterClass(alwaysRun = true)
